requirements.txt
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
google-api-core==1.11.1
google-auth==1.6.3
google-cloud-language==1.2.0
googleapis-common-protos==1.6.0
grpcio==1.21.1
idna==2.8
mkl-fft==1.0.12
mkl-random==1.0.2
mysql-connector==2.2.9
numpy==1.16.4
pandas==0.24.2
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
rsa==4.0
selenium==3.141.0
six==1.12.0
urllib3==1.25.3
wincertstore==0.2
